Add extra ScheinStatusBox component for StudentInfoView.
