+++ client/src/view/studentmanagement/student-info/StudentInfo.tsx
@@ -1,6 +1,5 @@
-import { Box, CircularProgress, Typography } from '@material-ui/core';
-import { createStyles, makeStyles, useTheme } from '@material-ui/core/styles';
-import clsx from 'clsx';
+import { Box, Typography } from '@material-ui/core';
+import { createStyles, makeStyles } from '@material-ui/core/styles';
 import { useSnackbar } from 'notistack';
 import React, { useEffect, useState } from 'react';
 import { useParams } from 'react-router';
@@ -14,21 +13,13 @@ import { useErrorSnackbar } from '../../../hooks/useErrorSnackbar';
 import { getStudentOverviewPath } from '../../../routes/Routing.helpers';
 import CriteriaCharts from './components/CriteriaCharts';
 import EvaluationInformation from './components/EvaluationInformation';
+import ScheinStatusBox from './components/ScheinStatusBox';
 
 const useStyles = makeStyles(theme =>
   createStyles({
     backButton: {
       marginRight: theme.spacing(2),
       alignSelf: 'center',
-    },
-    statusLoadingSpinner: {
-      marginRight: theme.spacing(1),
-    },
-    greenFont: {
-      color: theme.palette.green.main,
-    },
-    redFont: {
-      color: theme.palette.red.main,
     },
     criteriaCharts: {
       marginBottom: theme.spacing(1),
@@ -46,7 +37,6 @@ interface RouteParams {
 
 function StudentInfo(): JSX.Element {
   const classes = useStyles();
-  const theme = useTheme();
 
   const { studentId, tutorialId } = useParams<RouteParams>();
 
@@ -84,39 +74,7 @@ function StudentInfo(): JSX.Element {
 
         <Typography variant='h4'>{student && getNameOfEntity(student)}</Typography>
 
-        <Box
-          display='flex'
-          alignItems='center'
-          textAlign='center'
-          border={1}
-          marginLeft='auto'
-          borderRadius={4}
-          borderColor={
-            scheinStatus
-              ? scheinStatus.passed
-                ? theme.palette.green.main
-                : theme.palette.red.main
-              : 'divider'
-          }
-          padding='5px 15px'
-        >
-          {scheinStatus ? (
-            <Typography
-              variant='button'
-              className={clsx({
-                [classes.greenFont]: scheinStatus.passed,
-                [classes.redFont]: !scheinStatus.passed,
-              })}
-            >
-              Schein {scheinStatus.passed ? 'bestanden' : 'nicht bestanden'}
-            </Typography>
-          ) : (
-            <>
-              <CircularProgress size='1rem' className={classes.statusLoadingSpinner} />
-              <Typography variant='button'>Lade Status...</Typography>
-            </>
-          )}
-        </Box>
+        <ScheinStatusBox scheinStatus={scheinStatus} marginLeft='auto' />
       </Box>
 
       <Placeholder
